Modelacion de cercania y Stemmer arreglados, Arry.Sort

+++ MoogleEngine/SearchItem.cs
@@ -2,7 +2,7 @@ namespace MoogleEngine;
 
 public class SearchItem
 {
-    public SearchItem(string title, string[] snippet, int[] Pos_Snippet, float score)
+    public SearchItem(string title, string[] snippet, int[] Pos_Snippet, double score)
     {
         this.Title = title;
         this.Snippet = snippet;
@@ -15,5 +15,5 @@ public class SearchItem
     public string[] Snippet { get; private set; }
     public int[] Pos_Snippet { get; private set; }
 
-    public float Score { get; set; }
+    public double Score { get; set; }
 }
